Change bigints for integers

+++ database/create_users_and_profiles_tables.sql
@@ -5,7 +5,7 @@
 DROP TABLE IF EXISTS users CASCADE;
 
 CREATE TABLE users (
-  id               bigserial       PRIMARY KEY,
+  id               serial          PRIMARY KEY,
   username         varchar(40)     NOT NULL UNIQUE,
   first_name       varchar(40)     NOT NULL,
   last_name        varchar(40)     NOT NULL,
@@ -19,7 +19,7 @@ CREATE TABLE users (
 DROP TABLE IF EXISTS profiles CASCADE;
 
 CREATE TABLE profiles (
-  user_id          bigint          REFERENCES users ON DELETE CASCADE UNIQUE,
+  user_id          integer         REFERENCES users ON DELETE CASCADE UNIQUE,
   bio              varchar(512),
   birth_date       timestamp,
   gender_identity  gender_identity
